requirements.txt
streamlit
PyPDF2
python-docx
scikit-learn
sentence-transformers
pandas